add varbase.transits
